# docs/swagger.yaml
basePath: /api
definitions:
  apperror.AppError:
    properties:
      code:
        type: integer
      error_code:
        description: Machine-readable error code
        type: string
      message:
        type: string
    type: object
...
  dto.CreateOrganizationRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      organization_type:
        enum:
        - company
        - store
        type: string
      parent_organization_id:
        type: string
    required:
    - name
    - organization_type
    type: object
  dto.CreatePermissionRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
  dto.CreateRoleApprovalRequest:
    properties:
      description:
        maxLength: 255
        type: string
      requested_level:
        description: Level 100 reserved for superadmin
        maximum: 99
        minimum: 0
        type: integer
      requested_role_name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - requested_level
    - requested_role_name
    type: object
  dto.CreateRoleRequest:
    properties:
      description:
        maxLength: 255
        type: string
      level:
        description: Level 100 reserved for superadmin
        maximum: 99
        minimum: 0
        type: integer
      name:
        maxLength: 50
        minLength: 3
        type: string
      predefined_name:
        description: 'NEW: Android-style name'
        maxLength: 50
        minLength: 3
        type: string
    required:
    - level
    - name
    - predefined_name
    type: object
  dto.CreateUserRequest:
    properties:
      email:
        example: new.user@example.com
        type: string
      password:
        example: strongpassword123
        minLength: 8
        type: string
      role_id:
        example: b1c2d3e4-f5g6-7890-1234-567890abcdef
        type: string
      username:
        example: newuser
        minLength: 3
        type: string
    required:
    - email
    - password
    - role_id
    - username
    type: object
...
  dto.JoinOrganizationRequest:
    properties:
      organization_code:
        type: string
    required:
    - organization_code
    type: object
  dto.LoginRequest:
    properties:
      password:
        example: password
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      message:
        example: Login successful
        type: string
      permissions:
        items:
          type: string
        type: array
      user:
        allOf:
        - $ref: '#/definitions/dto.UserResponse'
        description: Menggunakan UserResponse DTO
    type: object
...
  dto.OrganizationResponse:
    properties:
      child_organizations:
        items:
          $ref: '#/definitions/dto.OrganizationResponse'
        type: array
      code:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/dto.UserResponse'
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      member_count:
        type: integer
      name:
        type: string
      organization_type:
        type: string
      parent_organization:
        $ref: '#/definitions/dto.OrganizationResponse'
      parent_organization_id:
        type: string
      updated_at:
        type: string
    type: object
  dto.PagedUserResponse:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
      users:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
    type: object
  dto.PermissionResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.RefreshTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  dto.RoleResponse:
    properties:
      description:
        type: string
      id:
        type: string
      is_active:
        description: 'NEW: Role activation status'
        type: boolean
      is_system_role:
        description: 'NEW: System role protection'
        type: boolean
      level:
        description: 'NEW: Hierarchy level'
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      predefined_name:
        description: 'NEW: Android-style name'
        type: string
    type: object
...
  dto.UpdateOrganizationRequest:
    properties:
      description:
        maxLength: 500
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 3
        type: string
    type: object
...
  dto.UserOrganizationResponse:
    properties:
      is_active:
        type: boolean
      joined_at:
        type: string
      organization:
        $ref: '#/definitions/dto.OrganizationResponse'
      organization_id:
        type: string
      role_in_organization:
        type: string
      user_id:
        type: string
    type: object
  dto.UserResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.png
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        type: string
      role:
        example: user
        type: string
      role_id:
        example: b1c2d3e4-f5g6-7890-1234-567890abcdef
        type: string
      username:
        example: johndoe
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Beresin App backend.
  title: Beresin App API
  version: "1.0"
paths:
  /admin/organizations:
    post:
      consumes:
      - application/json
      description: Creates a new organization. Requires platform level permissions.
      parameters:
      - description: Organization Details
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Organization created successfully
          schema:
            $ref: '#/definitions/dto.OrganizationResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Create a new organization
      tags:
      - Admin
      - Organizations
  /admin/organizations/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an organization. Requires platform level permissions.
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Organization deleted successfully
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Delete organization
      tags:
      - Admin
      - Organizations
    put:
      consumes:
      - application/json
      description: Updates organization details. Requires admin permissions for the
        organization.
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Organization Details
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOrganizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Organization updated successfully
          schema:
            $ref: '#/definitions/dto.OrganizationResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Update organization
      tags:
      - Admin
      - Organizations
...
  /admin/permissions:
    get:
      consumes:
      - application/json
      description: Retrieves all available permissions in the system. Requires 'permissions:read'
        permission.
      produces:
      - application/json
      responses:
        "200":
          description: Permissions retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.PermissionResponse'
            type: array
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: List all permissions
      tags:
      - Admin
      - Permissions
    post:
      consumes:
      - application/json
      description: Creates a new permission in the system. Requires 'permissions:create'
        permission.
      parameters:
      - description: New Permission Details
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Permission created successfully
          schema:
            $ref: '#/definitions/dto.PermissionResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "409":
          description: Permission with that name already exists
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Create a new permission
      tags:
      - Admin
      - Permissions
...
  /admin/roles:
    get:
      consumes:
      - application/json
      description: Retrieves all roles with their permissions. Requires 'roles:read'
        permission.
      produces:
      - application/json
      responses:
        "200":
          description: Roles retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.RoleResponse'
            type: array
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: List all roles
      tags:
      - Admin
      - Roles
    post:
      consumes:
      - application/json
      description: Creates a new role. Requires 'roles:create' permission.
      parameters:
      - description: New Role Details
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role created successfully
          schema:
            $ref: '#/definitions/dto.RoleResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "409":
          description: Role with that name already exists
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - Admin
      - Roles
...
  /admin/users:
    get:
      description: Retrieves a paginated list of all users. Requires 'users:read'
        permission.
      parameters:
      - default: 1
        description: Page number for pagination
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page for pagination
        in: query
        name: limit
        type: integer
      - description: Search term for username or email
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: A paginated list of users
          schema:
            $ref: '#/definitions/dto.PagedUserResponse'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - Admin
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided details. Requires 'users:create'
        permission.
      parameters:
      - description: New User Details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/apperror.AppError'
        "409":
          description: Username or email already exists
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - Admin
      - Users
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns an access token. The refresh token
        is set in an HttpOnly cookie.
      parameters:
      - description: Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/apperror.AppError'
      summary: User Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidates the refresh token and clears the session cookie.
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Failed to read refresh token
          schema:
            $ref: '#/definitions/apperror.AppError'
      summary: Logout user
      tags:
      - Auth
  /auth/me:
    get:
      description: Returns the current authenticated user's information with their
        permissions.
      produces:
      - application/json
      responses:
        "200":
          description: Current user information
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get Current User Information
      tags:
      - Auth
  /auth/refresh:
    post:
      description: Generates a new access token using a valid refresh token from the
        cookie.
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/dto.RefreshTokenResponse'
        "401":
          description: Unauthorized or invalid refresh token
          schema:
            $ref: '#/definitions/apperror.AppError'
      summary: Refresh Access Token
      tags:
      - Auth
...
  /organizations:
    get:
      consumes:
      - application/json
      description: Retrieves organizations with optional filters
      parameters:
      - description: Organization Type
        enum:
        - platform
        - company
        - store
        in: query
        name: type
        type: string
      - description: Parent Organization ID
        in: query
        name: parent_id
        type: string
      - description: Is Active
        in: query
        name: active
        type: boolean
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Organizations retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.OrganizationResponse'
            type: array
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: List organizations
      tags:
      - Organizations
  /organizations/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves organization details by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization retrieved successfully
          schema:
            $ref: '#/definitions/dto.OrganizationResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get organization by ID
      tags:
      - Organizations
  /organizations/{id}/leave:
    delete:
      consumes:
      - application/json
      description: Allows a user to leave an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully left organization
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Leave organization
      tags:
      - Organizations
      - Users
  /organizations/code/{code}:
    get:
      consumes:
      - application/json
      description: Retrieves organization details by code
      parameters:
      - description: Organization Code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization retrieved successfully
          schema:
            $ref: '#/definitions/dto.OrganizationResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get organization by code
      tags:
      - Organizations
  /organizations/join:
    post:
      consumes:
      - application/json
      description: Allows a user to join an organization using organization code
      parameters:
      - description: Join Organization Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JoinOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully joined organization
          schema:
            $ref: '#/definitions/dto.UserOrganizationResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/apperror.AppError'
        "409":
          description: User already member of organization
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Join organization
      tags:
      - Organizations
      - Users
  /roles/approval-requests:
    post:
      consumes:
      - application/json
      description: Creates a new role approval request. Users can request new roles
        to be created.
      parameters:
      - description: Role Approval Request Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleApprovalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role approval request created successfully
          schema:
            $ref: '#/definitions/dto.RoleApprovalResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.AppError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Create a role approval request
      tags:
      - Roles
      - Approval
  /roles/predefined-options:
    get:
      consumes:
      - application/json
      description: Retrieves available predefined role options with their levels and
        descriptions.
      produces:
      - application/json
      responses:
        "200":
          description: Predefined role options retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.PredefinedRoleOption'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get predefined role options
      tags:
      - Roles
...
  /users/me/organizations:
    get:
      consumes:
      - application/json
      description: Retrieves all organizations for the current user
      produces:
      - application/json
      responses:
        "200":
          description: User organizations retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.UserOrganizationResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get user organizations
      tags:
      - Organizations
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"